feat(cli): sync config-lifecycle response models with the gateway contract
The persist and restore responses grew identity, coverage and durability
fields that the CLI models did not carry: the persisted document's schema
version and lineage generation, the domains a document captures and the ones
it deliberately never captures, the recovery-dependency manifest with each
entry's disposition, and - for a committed restore - whether the applied
configuration was actually written through to disk.

The new fields are pointers and slices so "the gateway did not report it"
stays distinguishable from "the gateway reported a zero". An older gateway
answering the same endpoints is reported as speaking the legacy contract, and
the CLI states what it was not told instead of printing a zero as if it were
an answer.

Alongside the models this adds the pieces the commands need to be honest
about a result:

- stable reason codes for every failure class, so automation branches on a
  code rather than on prose, plus a shared JSON envelope carrying the code,
  the message, the HTTP status and the response body behind the verdict;
- verdict rules read from the body rather than the status code, because the
  restore endpoint answers 200, 400 and 500 with the same result document.
  A committed restore whose write-through failed is a failure: it applied,
  but the configuration will not survive the next restart;
- snapshot checksum verification. The response body is the document's
  canonical encoding, so blanking the checksum value reproduces exactly the
  bytes the gateway hashed - no re-encoding, and no second owner of the
  schema. A document with no checksum comes from an older gateway and is
  reported as unverified, never as verified;
- an atomic snapshot store: temporary file in the destination directory,
  flushed, renamed into place, 0600. A failed download can no longer destroy
  the snapshot already sitting at that path.

The API server port is widened from int16 to int in the same change. Half the
port space - everything above 32767, which includes the entire ephemeral
range - could not be expressed, so a gateway listening there was unreachable
from the CLI.

+++ pkg/api/rest.go
@@ -65,7 +65,10 @@ type RESTOptions struct {
 	PrintOption string
 	Protocol    string
 	ServerIP    string
-	ServerPort  int16
+	// ServerPort is a full TCP port number. It is deliberately not an
+	// int16: half the port space (32768-65535, which includes the
+	// ephemeral range a test or a port-forward lands in) does not fit one.
+	ServerPort  int
 	Timeout     int16
 	ServiceName string
 	Token       string
@@ -90,7 +93,7 @@ func (r *RESTClient) GetProcotol() string {
 }
 
 func (r *RESTClient) GetHost() string {
-	return fmt.Sprintf("%s:%d", r.Options.ServerIP, int(r.Options.ServerPort))
+	return fmt.Sprintf("%s:%d", r.Options.ServerIP, r.Options.ServerPort)
 }
 
 func (r *RESTClient) GET(ctx context.Context, getURL string) (*http.Response, error) {
